+++ sdk/nodejs/runtime/index.ts
@@ -1,6 +1,6 @@
 // Copyright 2016-2017, Pulumi Corporation.  All rights reserved.
 
-export * from "./closureSerialization";
+export { serializeFunctionAsync }  from "./closure/serializeClosure";
 export * from "./config";
 export * from "./invoke";
 export * from "./resource";
