+++ sdk/nodejs/tsconfig.json
@@ -32,10 +32,10 @@
         "log/index.ts",
 
         "runtime/index.ts",
-        "runtime/closureCreation.ts",
-        "runtime/closureFunction.ts",
-        "runtime/closureSerialization.ts",
-        "runtime/closureSuper.ts",
+        "runtime/closure/createClosure.ts",
+        "runtime/closure/parseFunction.ts",
+        "runtime/closure/serializeClosure.ts",
+        "runtime/closure/rewriteSuper.ts",
         "runtime/config.ts",
         "runtime/debuggable.ts",
         "runtime/invoke.ts",
